move modal provider to own file
