+++ src/index.jsx
@@ -1,46 +1,9 @@
 import React, { Component } from 'react'
 import ReactDOM from 'react-dom'
 import styled, { css } from 'styled-components'
-import { Provider, Consumer } from './context'
+import { Consumer } from './context'
+import ModalProvider from './ModalProvider'
 import { BaseModalBackground } from './baseStyles'
-
-class ModalProvider extends Component {
-  constructor (props) {
-    super(props)
-
-    this.state = {
-      modalNode: null,
-      BackgroundComponent: BaseModalBackground
-    }
-
-    this.setModalNode = this.setModalNode.bind(this)
-  }
-
-  static getDerivedStateFromProps (nextProps, prevState) {
-    if (nextProps.backgroundComponent !== prevState.BackgroundComponent &&
-        nextProps.backgroundComponent) {
-      return { BackgroundComponent: nextProps.backgroundComponent }
-    }
-
-    return null
-  }
-
-  setModalNode (node) {
-    this.setState({ modalNode: node })
-  }
-
-  render () {
-    return (
-      <Provider value={{
-        modalNode: this.state.modalNode,
-        BackgroundComponent: this.state.BackgroundComponent
-      }}>
-        {this.props.children}
-        <div ref={this.setModalNode} />
-      </Provider>
-    )
-  }
-}
 
 class Modal extends Component {
   constructor (props) {
